Add a player character (frog) and jumping
This will be a one button game. The plan is to allow the player to vary
jumps by holding down the button for different lengths of time, but for
now, just have a constant jump. Jump in the direction the player is
facing, and stick to walls and (untested) ceilings.

Export jump strength and fall acceleration for adjustment from the
editor.

Add a public domain spritesheet (from which the project icon was taken).
Animate an idle blinking animation, as well as jumping and landing.

TODO: Test landing on the ceiling.