Use rotation to place player on the ceiling
When we simply flip the sprite, as previously done, the hit box is
unchanged. As currently written, this positioned the player so they look
like they are too far into the ceiling. When we rotate the entire
Player, the hit box still lines up as desired, so the player looks to
be sitting on the ceiling, just like when on the floor.