Fix direction for jumps from the ceiling
Keep track of whether the player is on the ceiling or not. As with being
on the floor, using `CharacterBody2D::is_on_ceiling` is not working as I
expected.

Create the entire jump vector in a helper function. This keeps the
handling of being on the ceiling together, since it affects both jump
direction and rotation. (I could have left this as-is and used the angle
to rotate all the way from up to down at 45 degrees, but this seemed
more intuitive.)

TODO: When jumping off of the ceiling, the jump animation should start
upside down, and only flip the frog on the last frame.