Push off against the ceiling upside down
While it looks natural to flip right side up when jumping off the walls,
it is jarring to see when jumping off the ceiling. Leave the orientation
unchanged for the jump, so it looks like the player is pushing off the
ceiling.

Another thing I tried - push off while still upside, but then flip right
side up at the end of the animation. This did not look right to me.

When landing on a floor, ensure that the orientation is correct. Handle
the final case of a mysterious collision normal ({0, 0}?) with an error
message.