Modulate jump by length of button press
Introduce `JumpHandler` to handle input and determine the strength of a
jump based on how long the button was held. Make `JumpHandler` inherit
from `Node` so it can have attributes that can be modified by the
editor.

TODO: Add visual indicator of jump strength.
TODO: Remove log statements for jumping.