Add `JumpMeter`
This custom `Control` allows me to show a progress bar/meter displaying
how strong the jump will be when the button is released.

Export many fields to the editor in order to fine tune.

TODO: Move so it's always displayed below the player.
TODO: Split this into an add-on so it can be used in other projects?