Consistently draw the jump meter below the player
`player.rs` rotates the entire `CharacterBody2D`, and the `JumpMeter` is
a child of that node, so it gets rotated, too. Undo the matrix and draw
using global coordinates so that it always shows below the player.

FIXME: `offset_y` isn't being exported to the editor properly. I'm
wondering if it's because I already have too many exported fields?
This is the ninth one, FWIW.