Add a dragonfly to eat
Add a new asset for the dragonfly, along with info about its source and
license.

Set up a simple script for the dragonfly, so it disappears when the
player touches it.

Add the player to the (new) `player` group so that the dragonfly can
check what body entered. (Perhaps redundant with collision layers, but
just in case we need it later.)