Simplify dragonfly animation
I believe the first four are the "idle" animation. I cannot tell the
difference between it and the next four frames, which I believe
represent "movement". But taking out the last four, "damage" makes it
look better. ("damage" might be fine as a separate animation, but this
game does not require it.)