Allow specifying whether the player starts on a surface
In the main scene, the player starts off in mid-air, and the animation
starts on the falling frame. For the tutorial, I want the player to
start on a surface, so allow the editor to specify so. If the player is
on a surface, use the frame representing sitting still.