`JumpHandler`: use Godot time instead of system time
Relying on the system time resulted in inconsistent output. Attempting
to hold the button down for the same amount of time resulted in
variable button presses. This was noticeable in the tutorial, where some
initial jumps ended up longer than others.

Using the cumulative time from the provided `delta` updates seems to
result in a more consistent jump.

Note that the tutorial still uses system time. If it proves to still be
inconsistent, this may be worth modifying as well. (Also note that a
possible issue is that system time continues to pass between the
tutorial's read of the time and the jump handler's.)