Improve determination of landing direction
Prior code assumed that the normal of a collision would point directly
along an axis. In practice, sometimes the normal takes a form like
`(-0.0050424966, -0.99998724)`. In the old code, the x component being
less than zero resulted in rotating the player as if they landed on the
left side of a surface. In reality, this should be treated as landing
on top of a surface.

Occasionally, the old code resulted in the player landing with the wrong
rotation. I have only successfully reproduced the case above, with a
very small jump.

Fix the code by increasing the tolerance of the comparisons. (Note:
using `Vector2::round` and leaving the comparisons alone would have had
a similar effect.) Testing reveals that this fixes the one case I could
reproduce.