Support touch input
Given that this only requires one button(-type input), it seems
reasonable to make it mobile friendly, where a screen press is used for
input. Add a new `JumpDetector` implementation that reads screen
touches.

In `player.rs`, check for touch input. If the "jump" input action has
not been used, switch to touch input.

In `tutorial.rs`, load the main level on touch input (release), as with
a key press. Unlike key input (TODO: update to match?), load on release,
in order to prevent awkward handling on the level. (e.g. player is
touching on load, but that gets missed because it started before the
level loaded.)

Emulate touch from the mouse for testing. TODO: Support mouse input for
real?