Switch the stretch mode to "viewport"
In the project settings, switch the stretch mode to "viewport". This
stretches out the game so that it fills the screen.

Note that it seems the "Canvas Resize Policy" in the export options
must be "Adaptive". (This might be the default? it's how I initially
created the file, but at some point along the way I tried changing it to
"Project", thinking it would respect my stretch mode, which is part of
the project settings. But I suspect it is specifically respecting the
width and height.)

Now it has a good size on an iPad mini. It is better on my phone, but
still a little small. With the current dimensions, it might be better in
landscape.

TODO: If I intend to make playing on a phone a major use case, I should
change the dimensions.