Remove the "wall" group
This was used to identify a surface the frog can land on. One downside
to using this group was that when I create a new level, I forget to add
the group, and things don't work. If I stored the `TileMapLayer` as a
scene, I think I could save the group with the scene, but more broadly,
I think this is better handled using physics layers. The `TileSet`
already has the proper physics layer set, so the "wall" group is
redundant.