Add an alligator enemy with a couple of animations
The alligator can be tested by themself, but you need to modify the
transform so that they're on screen.

Set up some animations: blinking and shifting eyes for idle, and
flashing eyes for when they are focused on the player.

Add some keys for testing out the animations. TODO: Remove