WIP: Animate eating the player
Set the alligator's z-index in front of the fake player, so it looks
like the player is inside the alligator's mouth. TODO: This is only set
in the test scene; I'll need to modify more generally, while keeping
other z-indices in mind.

Provide a state machine for the alligator, so we can animate opening and
closing the alligator's jaw.

TODO: Remove the player
TODO: Animate the player being eaten?
TODO: Make the alligator gloat