Move the "fake" player into a standalone scene
My first attempt to continue testing in the `test_alligator` scene was
to reuse the "FakePlayer", but this had surprising behavior: the object
triggered collisions with the alligator's `Area2D` regions when being
re-added to the tree, even if its current position did not intersect the
region.

Instead, we'll create a new fake player the next time one is needed.
This commit simply factors out a new `tscn` file, so that a follow-on
commit can instantiate from a `PackedScene`. This commit does not change
any behavior.