Add another `Area2D` to effect eating the player
It didn't look right in my attempts to use simple math. Now if the
player lands in the right part of the jaw, the alligator will eat the
player.

Increase the alligator's z-index so it draws in front of the player.
This makes it look more like the frog is in the alligator's mouth.

Close the alligator's mouth if the frog escapes unscathed.