Add a tile with smaller lily pads
Originally I had open water surrounding the alligator, but this allows
the frog to fall through[^1]. Swimming as the player would require a
whole new mechanic, and falling into water shouldn't be fatal for a frog
(though I've seen this in other games). Smaller lily pads provide a
landing surface while allowing an alligator to hide out in them.

[^1]: Another possible solution to consider later: the alligator swims
to reach the frog as necessary.