Add a level with an alligator and some level behavior
The level behavior primarily revolves around player death and
respawning, so include a level with an enemy.

Add a "predators" group for enemies that can eat the player. (Place the
alligator in it.) All predators should have a "player_eaten" signal,
which `level.rs` connects to its own callback.

Track the player's orientation at level start (`PlayerInfo`) so that it
can be used to respawn the player later. An alternative would've been
to use exported variables directly in `Level`, but retrieving from the
player allows placing the player in the level using the editor in a
natural way. (As opposed to setting, in particular, the position, via
exported variables in the level.)

Make `Level` inherit from `TileMapLayer`, since that is the root of my
levels so far. This could be revisited if another root node makes more
sense.