Vary how often the player blinks
Rather than relying on the editor-specified wait time and autostart,
follow the model already used by the alligator: use some randomness to
determine when blinks happen.

This is more relevant in a level with multiple frogs - it prevents them
(almost[^1]) always blinking in unison.

[^1]: A frog doesn't blink while jumping, but any frogs that were
stationary on the 5-second interval would simultaneously blink.