NOT FOR SUBMISSION: Print `TileData` for collision
Again, I'm not sure I want this, but it might be useful to have a
cleaned-up version that I can turn on and off as desired.

Regardless, I learned some useful information. I can use the global
position (which `TileMapLayer`, via `Node2D`, first converts to local
coordinates) to find out which tile had the collision. From the
`TileData`, I can get the polygons used for physics collisions. They
are a bit mysterious, but the coordinates are local to the tile, with
(0,0) in the middle. (I'm using 64x64 tiles, so the edge is +/-32.)
`TileMapLayer` can also let me know the local (then convert to global)
coordinates of the tile, so I can find the global coordinates of the
shape. Note that the raw coordinates match what's in the tile set data,
which makes sense.

And one of the coordinates matches the collision position exactly! Now
to figure out how to adjust the player in response...