Player: move offset from collision rect to sprite
The goal of the offset in the collision rect was to better line up
with the sprite. But I think the math for positioning the player will
be simpler if the collision rect is *not* offset, and offsetting the
sprite (in reverse so they align) ought to not have any negative impact.
Note that the sprite previously had a very tiny offset that is not
noticeable (and I don't remember how it might have gotten set that way),
which I've left out of the new setting.

I have not yet attempted to adjust the jump meter to make up for this
change, but at least when the player is right-side up and facing right,
it looks fine. It might even be better aligned. If it does need to be
adjusted, that would happen in the script.