Handle being on the left side of a tree better
Previously, we rotated the `Player` to the right and used `set_flip_v`
on the sprite to position the player. But this resulted in a
misalignment of the collision rectangle and the sprite. Rotate in the
other direction, so they align better. Now we never need to
`set_flip_v`.