WIP: Manually place the player on landing
Start off with landing directly on the corner, which inspired this
branch. Pick an adjacent edge which has enough room for the player.

TODO: If both have enough room, pick one based on player's motion,
probably?
TODO: If neither have enough room, then what? Maybe reduce the required
space to begin with. If there truly isn't enough room, maybe we could
keep looking at nearby edges?

Compute the normal based on the chosen side's corners and the player's
incoming motion. Use this directly as the player's rotation. This is
simpler than the old method, and it's more flexible. It ought to support
ramps (untested), and the player will line up better even on an
accidental slight incline. This should also make it easier to avoid
collisions with the surface.

(IDEA! TODO: After rotation and manually positioning the player, maybe
we should run `move_and_collide` again? Whether we just test (and report
the situation) or try to handle it, this might be good? The levels
should be designed such that the player wouldn't collide elsewhere,
anyhow.)

Fallback to the old calculation for rotation in collisions I have not
yet updated.

TODO: Handle shapes that align with adjacent cells to make a bigger
landing spot.
TODO: Handle hitting an edge directly, rather than a corner.
TODO: Handle narrow edges (e.g. the bottom of the vine)
TODO: Reposition the player.
TODO? Handle surfaces with slight angles in them.

Run `rustfmt`.