During testing, allow restarting/changing levels
During testing, it's nice to be able to restart a level without going
back to the editor. Add the notion of a test level. During testing, the
user has access to more functions, like restarting the level with
ESCAPE. Make `test_corner` a test level.

In addition, I am imagining a series of test levels that the user can
cycle through quickly. To facilitate that, add a field for the next
level. During testing, the user can hit ENTER to load the next level.
This exported variable could also be used for real levels, to be
loaded when the player completes the level.