WIP: Position from the corner
When hitting a corner, adjust the player's postion along the edge so
that the player fits on the edge. Use the normal to adjust away from
the surface, and double check for new collisions.

The player might be adjusted too much in one frame, though. Will revisit
this later, after other collisions look good.