Refactor `pick_side_to_land_on` for wider use
Append `from_corner` to the end to since I plan to add other functions
that behave similarly, but not from a corner.

Pass in the index, instead of the collision. This feels a little more
natural to me, since we had to search the array to find that the
collision is in it. A future commit will treat a corner as the
collision location, even though it's not, so it feels more natural to
pass the index. Though I'll admit that I then have to find the corner
again, so it's probably fine either way.