Adjust player position to parent's local position
When computing the player's final position, consider that the parent
may not be at (0,0). This is the case in `level.tscn`, leading to
the wrong position.

Note that `PlayerInfo.pos` is local to the parent.

The only other call to `set_position` uses an offset, rather than a
global position, so it doesn't need to take the level's coordinates into
account.