WIP: Use normals to determine side to land on
When landing on a corner, we previously picked the "prior" side (i.e.
the side containing the prior point in the polygon) if it's large
enough. That's not quite right though. Choose the side to try first
based on the collision normal, and still fall back to the other side if
there's not enough room on the side where the collision occurred.

Perhaps this is a workaround for `test_corner_seam`; if the player lands
on a seam like this, it should not be treated as a corner. However, this
still helps find the proper side.

TODO: No need to push away from the corner since it seams with another
      tile.