Fix the bug in `test_edge.tscn`
If the player lands *close* to a corner, the collision may appear to be
in the plane of the adjacent side, resulting in picking the wrong
surface to "land" on, resulting in the player having the wrong
orientation.

Check the normal to verify that this is the correct side. Note that this
means `pick_side_to_land_on` and `pick_side_to_land_on_from_corner`
share more behavior (comparing normals), so there might be a clean way
to merge them.