Align player with landing surface
This was already handled for a corner collision, since the position
is moved away from the corner, directly on the surface.

When landing directly on an edge, changing the rotation means that the
player may no longer be aligned with the surface. Adjust the player to
be resting directly on the surface.

Add a print statement for the player's position after landing. This is
helpful for setting up repro cases.

Update `test_corner_seam.tscn` to note a repro case that uncovered this
bug. Landing on an edge was using the collision location as the basis
for placing the player, but it acted as though the collision was with
the center of the player. It looks smoother to just adjust in the
direction of the normal.