Add a test and fix landing on a corner that seams up
Follow on to 2322334aedaba51294165e57315e123c5f44a757. Landing on a
corner that seams with the tile above should behave the same.

Add a test case, and insert it into our chain of test scenes, which
currently looks like

`test_edge` -> `test_corner_seam_left` -> `test_corner_seam_up` ->
    `test_corner` -> `test_alligator`

Note that this code now also supports tiles that seam right or down,
though I have not been able to reproduce these cases. It may be that
there is some bias such that it never happens. Add error statements so
we can easily spot cases when this happens so we can add test cases in
the future.