Fix landing on an almost 180 degree corner
If two edges are at an angle close to 180 degrees, treat them as a
single surface. For smaller angles, the player would move away from the
corner, but that isn't necessary here.

Remove the variable `hit_a_corner`, and special case moving away from a
corner based on the `LandingSurface` instead, since we treat this case
not hitting a corner even though technically the player did.

Export a new variable from `jump_handler.rs`, to set a maximum jump. And
add a print statement displaying the current strength. This makes it
easier to create a repro case. Add a repro case for this bug.

New chain for test scenes looks like

    `test_slant` -> `test_edge` -> `test_corner_seam_left` ->
	`test_corner_seam_up` -> `test_corner` -> `test_alligator`