WIP: When landing on a corner, shimmy into place
Rather than warping to a place where the player fully fits on the
surface, animate the player to the desired location.

Add a new animation, using frames from the jump animation but without
extending the legs all the way. This makes it look like the frog is
scrambling (or shimmying) to get fully onto the surface.

TODO: This works in `test_corner`, but we need more tests to verify.
FIXME: In `level`, we never stop shimmying.
FIXME: In `test_nook` (not yet checked in), we never stop shimmying
after a second jump.