Check for adjacent tiles even if collision tile is empty
In the new test scene, the collision of a full strength jump is directly
on the tile boundary. `TileMapLayer::local_to_map` for this point
returns the empty tile below. We want the tile above. We already have
code that checks for the tile above; use it even if though the first
tile was empty.

New test chain:

    `test_empty_tile` -> `test_nook` -> `test_corner_angle` ->
 		`test_slant` -> `test_edge` ->
                `test_corner_seam_left` -> `test_corner_seam_up` ->
                `test_corner` -> `test_alligator`