Fix new bug shimmying on tile edge
Previous commit introduced a bug where if the player lands near the edge
of a tile, it looks like they are hanging off the edge of a surface, so
they shimmy on. But if the surface continues onto the next tile, they
should stop in place. Augment `get_collider_points_from_tile_map_layer`
to catch such a case.

`test_corner_seam_up` contains a repro case. Add to its notes
documenting this.