Account for offset in `find_shimmy_dest`
In the new test scene, `test_tree_top`, the `TileMapLayer` is offset,
revealing that `find_shimmy_dest` didn't account for it. Add new helper
methods to repurpose existing code for converting between global and
local positions.

TODO: Combine the corner case with the edge case?
FIXME: In `test_slant` the player may shimmy backwards, seemingly only
if the player does *not* land directly on the corner.