Avoid more unnecessary shimmying
In this case, the player lands near a corner that doesn't quite line up
with the tile next to it. Check for the points being really close
together and treat them as a single point.

TODO: Share code with `pick_adjacent_side`?

Add a test. Note that while it looks similar to `test_edge`, that scene
does not exercise this case.