Consolidate checks for adjacent points
Sometimes tiles do not seam perfectly (and getting them to is tedious).
Previously we checked in three separate places for points that are
*almost* the same. Consolidate into a single place when returning the
points for the collider. This is unnecessary work, in cases where the
player collides with an unrelated side. If it were to show any perf
problems, perhaps we could store the real polygon on the `TileData` as
custom data.

Existing test cases pass.