More test scenes, including an intro
Add an intro test scene with instructions for running through the tests.
In addition to providing straightforward instructions where they are
relevant, this means that I can always start testing with the same
scene (`test_bed`), rather than having to remember what was the last
scene added at the beginning.

TODO: It'd be fun to make the frogs jump on their own in `test_bed`.

Add two scenes related to landing on the branch tip.

In `test_branch_tip`, the player lands on the short side of the branch.
TODO: Decide how to resolve this and fix.

In `test_branch_tip2`, the player lands directly on the corner, and the
current code computes the normal incorrectly, placing the player
overlapping with the branch. Fix forthcoming.