Fix `test_branch_tip2` case
We picked the normal for the `LandingSurface` based on the player's
incoming motion. In the test scene, the player is moving down and left,
and they collide with the bottom right corner. Because the right side,
the tip of the branch, is too short, the player should land on the
bottom. But the motion vector determines that the player should be
landing on top of this side. That is the interior of the branch,
resulting in a new collision, which warps the player outside the branch
in the next frame.

When landing on a corner, make an extra check to verify the normal
points in the proper direction.

TODO: Continue to think about a more robust way to do this. Note that
the player's position may not work either; in this case, the player is
above the bottom surface of the branch, too.