Add test for landing on corner of a too small side
I'm currently working on fixing a case where the player lands on a tiny
edge - we should act like they landed on a longer edge. (I have the code
more or less working, but it seems to have either introduced or exposed
some other bugs...) That bug is demonstrated in `test_branch_tip`.

This one is similar, only the player lands directly on the corner, which
follows a different code path. It seems I'm not taking into account the
adjacent tile which makes the surface longer. Once that is taken into
account, I expect this to work.