Increase tolerance for determining `LandingSurface`
When colliding, Godot tells us the object that was collided with, but it
doesn't specify which side of the object resulted in the collision. We
determine this by looking through the list of sides and finding one with
approximately the right normal and a small distance to the collision.

In the new test scene, `test_no_surface`, the collision is about 0.163
away from the plane of the landing surface. This doesn't meet the old
TOLERANCE of 0.05, so we reported no landing surface. Increase it to
a bigger round number, 0.2, so the player lands properly on the surface
and is lined up with it.

TODO: Perhaps it would be even better to check the distance to the line
segment, rather than just the plane. If the polygon looked something
like

```
____
|   |a
| |-
| |
| |_
|   |b
|---
```
the player might land on `b`, but we could pick `a` instead. Currently,
the code lines up the player from their current position onto the plane,
which should usually work out fine. e.g. if `a` and `b` are on parallel
lines, this will still work out. And if they're close, it should still
look ok. If they're *not* close, the distance of the player to the plane
won't be either.