Add docs comments to `jump_handler.rs`
Since I just set a range on `max_jump_strength_for_testing`, it seems
reasonable to also provide a comment.

Unfortunately, it is *not* showing up in the editor. Somewhere I think I
read you may have to provide docs comments on prior variables in the
same struct, so go ahead and write docs comments for them, too.

Note that this was working in 275a2a9b8b2655f160e450db9b1617257026b4e6 [^1].
But that same variable's docs are not currently showing in the editor
either.

[^1]: Where prior variables have no docs comments either.