Fix small jumps that do not leave a surface
When jumping off a wall, the player's orientation changes by 90 degrees.
Since the player's bounding box is wider than it is tall, this means
they may re-collide with the surface they're attempting to leave. This
is more pronounced with small jumps, since a larger jump may move them
out of range in the first frame.

Account for the change in size and move the player away from the wall
so they can continue the jump.

Add the test scene `test_edge`.