When landing on a small surface, try adjacent surfaces
If the surface is too small, look at the next and prior surfaces.

Add in a "safe margin" to the distance the player should rest off a
surface. Otherwise, moving around the corner has resulted in collisions,
preventing the player from shimmying onto the surface. Consolidate
calculation for this distance into a method.

TODO: Solve for the same case when hitting the corner