Add a test for another case of no landing surface
I discovered this after `test_no_surface` was fixed, so I think it
deserves its own test case. That said, this was incidentally fixed, I
believe since introducing `test_branch_tip3`, in
09dcbb3609e962c87667807b6a1e0c55c4cbcbdb.

Note: Previously I seemed to generate an extra collision at the start,
but that turned out to be due to forgetting to set the `Direction`
properly. This is now fixed.