WIP: Add a test scene for a weird bug
A full strength jump followed by another full strength jump results in
the player landing above the top-most branch and then warping below it.
(Further, they are sort of floating upside down near the trunk?)

I suspect the problem is that it has the right normal and distance from
plane to player, so I'll need to check that the player is in between the
two points.

But I don't understand why the recent fix seems to trigger this bug?
Maybe I just didn't notice the bug before?

TODO: Fix the problem
TODO: Start the player in position for the second jump?