Fix calls to `get_collision_polygon_points`
When the calling method, `get_collider_points_from_tile_map_layer`, was
originally written, `get_collision_polygons_count` was (pretty much)
always returning `1`, since the player collided with that tile. But now
that we check tiles adjacent to the collision [^1], we can easily query
a tile that has data but no polygons. This happens in the tutorial.
Prior to this commit, that resulted in an error message from the engine.

Check the count before querying for polygons, and skip the query if
none.

[^1]: For two reasons, which should show up in the git log. a) To build
a longer `LandingSurface` when a surface goes beyond the tile boundary.
b) If the collision is directly on a tile edge, we may start with the
wrong tile.