Fix warping in `test_warp`
This fixes the `TODO` in a3a81004adbc83ce7c5ebabee15170dd37acf80f. In
that commit, I suggest that the theoretical bug shouldn't be a problem,
but `test_warp` demonstrates a specific case where it is a problem. In
that scene, the first vertical side we test is too small[^1]. Had the
player fit on the side, they likely would've lined up properly where
they actually landed, but they end up being placed on the branch
sticking out from the trunk.

Fix this by checking the distance to both ends of the surface. We
already know the collision is roughly in the plane of the surface. If it
is also between the endpoints, the distances should add up to the full
surface length. This ensures we use the side where the player actually
landed.

Move the two existing checks to pick the right segment (previously in
`player.rs`'s `pick_side_to_land_on`) into a new method on
`LandingSurface`, `hit_by`, unchanged, except to add the distance check.

[^1] Only because of the tile boundary. We check adjacent tiles in some
cases, but to extend this surface longer would mean extending two
(rather than one) tiles away from the actual collision, which more
generally is not necessary.