Fix orientation in `test_nook`
When landing on a wall, the player's head is facing up. But the starting
code, focused on a player with no rotation, flips the sprite if the
player is facing right. In `test_nook`, this means they face down. Check
for a rotation and leave the player unflipped if it's non-zero. A more
thorough check might check for e.g. 45 degrees, but no scenes are set up
like that.