Add a doc comment to `JumpHandler`
A [theory was proposed](https://discord.com/channels/723850269347283004/1346579717280632863/1351459899443511297) that
the `struct` needs to be documented for property descriptions to show up
in the editor, so add a comment here.

As of right now, it works! After a `cargo build`, the editor now shows
descriptions for both `max_time_ms` and `max_jump_strength_for_testing`.

(FWIW, the editor still shows "*No description available.*" for the
`JumpHandler` itself, but at least the properties now work!)