Create a win condition
The player beats a level when they eat all of the prey.

Add `DragonFly` to a new group named "prey". This allows `Level` to
connect its new `prey_eaten` signal to a new method, so `Level` can
know when the player has eaten all the prey.

Add a message reporting victory when the player has eaten all the prey.

When the player has won, future actions will start the next level.