Add a new level that scrolls with a camera
This level is a work-in-progress[^1], but it demonstrates having a
camera that follows the player as they move through a larger world.
When the player dies, leave the camera in place, and restore it to the
player on respawn. Adjust the "WinMessage" so it is centered in front of
the camera.

Add to `level.tscn`'s "WinMessage" to instruct how to get to the next
level.

Modify ordering and collisions for better aesthetics. Place the floating
leaves in front of the player, so the player looks like they're in the
leaves, rather than in front of them. (Change the leaves' collision rect
so the player rests in the right spot.) Move the jump meter forward so
it is in front of the leaves, even though the player is drawn behind.

When beating the new level, a jump will return to the tutorial to play
again.

FIXME: It is easy to skip through the tutorial/start a jump on the next
level from the downpress on the finished level.

[^1]: As-is, I don't think this level best accounts for the player's
movement. It is easy to end up facing the wrong direction, and you have
to go a ways to get turned around and headed back up.