Turn on camera's position smoothing
This makes player warping less jarring when there is a camera following
the player.

Add a test scene which is a modified version of `scroll_level`. Compared
to the real scene, remove the "WinMessage", so there is no longer a win
condition. Perhaps this scene should be simplified further, but this
works fine, too.

For simplicity, leave it on always. To my eyes this looks fine overall,
and it improves warping a bunch.