Prevent player actions from affecting next scene
Currently, finishing `scroll_level` and pressing "jump" returns back to
the tutorial. Previously, that "jump" press would skip the tutorial and
(when using the keyboard) start filling the jump meter.

Make it so the player must press/tap again in order to jump/pass the
tutorial.

In the tutorial, move the check for "jump" action pressed from
`physics_process`, which would detect a key already held down, to
`unhandled_input`, where it only catches the initial press. (This is
also more consistent with `level`'s detection for respawning/starting
the next level.) Switch the touch check for the initial downpress as
well. These changes ensure that the player does not skip past the
tutorial on accident.

Similarly, make the `JumpHandler` check for the "jump" action being held
down at the start, preventing a press from the last level from also
starting to fill the jump meter. This requires changing the
`JumpDetector` trait, even though only action presses (vs mouse/touch
and the tutorial implementations) need this behavior.