WIP: Add the ability to trigger a bonus stage
Add an optional stage, set in the editor, to go to if the player eats a
bonus fly. If the stage is set, connect any "bonus_prey"'s "eaten"
signal to "on_bonus_found". When triggered, the next jump will lead to
the bonus stage.

TODO: Add a bonus message
TODO: Test