Replace "WinMessage" with separate scenes
Right now we have two messages and two scenes with win messages, one for
each, but separating these out makes it easier to share scenes and
modify in the editor.

Use a `CanvasLayer` as the root node, as suggested in the
[Dodge the Creeps tutorial](https://docs.godotengine.org/en/stable/getting_started/first_2d_game/06.heads_up_display.html).
This makes it easy to use an anchor for centering and avoids the need to
manually put the message in front of the camera.

Move `self.won = true;` to the top of the `if` block, since this is the
most significant game state change, in my mind.

Remove the old log statements if the message is missing. This should be
obvious.