Make a bonus level reachable from `vertical_level`
In `bonus_fly`, remove the `AnimatedSprite2D` from the "bonus_prey"
group, which it was previously added to accidentally. `level.rs` assumes
members in the group have an `eaten` signal, which is party of `Fly`
(which is also in the group).

In `vertical_level`, add a bonus fly and a bonus level.

Add a bonus message, shown programmatically when the bonus is found.

TODO: Integrate `vertical_level` with other levels.
TODO: Add regular flies.
TODO: Continue work on the bonus level, `swarm`.