Fix `player`'s appearance in the editor
Make `Player` a `tool` class, so we can run code in the editor. Move
code for picking the frame and `flip_h` to the beginning of `ready`, so
this code (and not the rest) can be run both in the editor and in
gameplay[^1].

Set for the default, too. In theory, this means I can change back to the
default and the `flip_h`/frame will change. Note that a scene may need
to be reloaded in order to see the change occur.

Skip `physics_process` (the only other method that I expect to be called
by the editor) when in the editor.

[^1]: To my surprise, the `flip_h` set in the editor's call to `ready`
does not affect gameplay, so we do need both.