Switch to programmatic win messages
Instead of setting the win message manually for each scene, switch to a
programmatic model as is done for the bonus message. If there is no next
scene listed, show the final level version. If there is, show the normal
version.

Remove the next level from the (current) final level. If there is no
next level specified, this is the final level, so switch back to the
tutorial.