Fix a shimmying bug in `scroll_level`
When jumping to a particular location, the player shimmies down. It
turns out this was a bug in the `TileMapLayer` - there were two tiles
with collision detection overlapping, which was not intentional. Remove
the redundant one, and now there is no unnecessary shimmying there.

Add a test level, mainly for the purpose of documentation, since the
bug was in the level itself.

Note to self: `vertical_level` has some tiles that are deliberately
overlapping, but I have not seen issues there.