Remove outdated `rustflags`
According to [the book](https://godot-rust.github.io/book/toolchain/export-web.html#admonition-note-1),
these are no longer needed.

Note that one of the removed features, `bulk-memory`, *sounds like* it
might be related to [a compilation error](https://discord.com/channels/723850269347283004/1359248427711594738),
but it is not. That was fixed by using a newer version of emsdk, 3.1.74.