Remove `tool` annotation from `Player`
Partial revert of e9904a11ded5874b507305611752e4eb5989e2da. This was
nice when it worked (though it requires some extra manual work of
reloading scenes), but it correlated with the editor crashing several
times.

Note that the docs [^1][^2] warn about the potential for crashes,
especially when freeing nodes. I'm not attempting to free any nodes from
the editor, so I'm not sure what I'm doing wrong. Regardless, removing
the annotation seems to have eliminated the crashes.

[^1]: https://docs.godotengine.org/en/stable/tutorials/scripting/gdscript/gdscript_basics.html#doc-gdscript-tool-mode
[^2]: https://docs.godotengine.org/en/stable/tutorials/plugins/running_code_in_the_editor.html