Tie respawning to existence of "Player"
`Level` respawns a player if it has no child named "Player". This makes
sense if the level started off with a child named "Player", but other
levels would spawn a new player unnecessarily. This led to workarounds
in levels that didn't even have a way for the player to die.

`Level` already checks to see if it has a "Player" at the start. Make
`player_respawn_info` an `Option`, only set if there is a "Player" in
`ready`. This way levels that don't need respawning no longer need to
work around it.

Remove `Default` from `PlayerInfo`, since it is no longer needed.

Revert workarounds from levels that no longer need them:
- "`swarm`: reparent `Player` to work around a bug"
  - commit bce1d41babb9a5bcce0bf2ecfd73f335ab949147.
- "Add `Player` to `test_bed`"
  - commit 5abf16927913137f0126f474197cd73dd44991eb.